add protection for main and dev

+++ utilities/delete_old_branch.sh
@@ -3,6 +3,7 @@
 git for-each-ref --sort=-committerdate --format="%(refname:short) %(committerdate:unix)" refs/remotes/ | awk -v two_weeks_ago=$(date -d '2 weeks ago' +%s) '$2 <= two_weeks_ago {print $1}' | while read branch; do
     remote_branch=$(echo $branch | sed 's#^origin/##')
     echo "deleting branch "$remote_branch"..."
-    git push origin --delete "$remote_branch" 2>/dev/null
-
+    if [ "$remote_branch" != "main" ] && [ "$remote_branch" != "develop" ]; then
+        git push origin --delete "$remote_branch" 2>/dev/null
+    fi
 done
